Format logs as JSON for both Datadog loggers
