+++ scfw/loggers/dd_agent_logger.py
@@ -2,49 +2,17 @@
 Configures a logger for sending firewall logs to a local Datadog Agent.
 """
 
-import json
 import logging
 import os
 import socket
 
-import scfw
 from scfw.configure import DD_AGENT_PORT_VAR
 from scfw.logger import FirewallLogger
-from scfw.loggers.dd_logger import DDLogger
+from scfw.loggers.dd_logger import DDLogFormatter, DDLogger
 
 _log = logging.getLogger(__name__)
 
 _DD_LOG_NAME = "dd_agent_log"
-
-
-class _DDLogFormatter(logging.Formatter):
-    """
-    A custom JSON formatter for firewall logs.
-    """
-    DD_ENV = "dev"
-    DD_VERSION = scfw.__version__
-    DD_SERVICE = DD_SOURCE = "scfw"
-
-    def format(self, record) -> str:
-        """
-        Format a log record as a JSON string.
-
-        Args:
-            record: The log record to be formatted.
-        """
-        log_record = {
-            "source": self.DD_SOURCE,
-            "service": self.DD_SERVICE,
-            "version": self.DD_VERSION
-        }
-
-        env = os.getenv("DD_ENV")
-        log_record["env"] = env if env else self.DD_ENV
-
-        for key in {"action", "created", "ecosystem", "msg", "targets"}:
-            log_record[key] = record.__dict__[key]
-
-        return json.dumps(log_record) + '\n'
 
 
 class _DDLogHandler(logging.Handler):
@@ -72,7 +40,7 @@ class _DDLogHandler(logging.Handler):
 
 # Configure a single logging handle for all `DDAgentLogger` instances to share
 _handler = _DDLogHandler() if os.getenv(DD_AGENT_PORT_VAR) else logging.NullHandler()
-_handler.setFormatter(_DDLogFormatter())
+_handler.setFormatter(DDLogFormatter())
 
 _ddlog = logging.getLogger(_DD_LOG_NAME)
 _ddlog.setLevel(logging.INFO)
